Refactor skeleton code

+++ src/main/java/kimjinung/platform/domain/order/OrderItem.java
@@ -27,7 +27,7 @@ public class OrderItem extends BaseEntity {
 
     private Integer quantity;
 
-    private Long orderPrice;
+    private Integer orderPrice;
 
     public OrderItem() {
     }
